+++ ost_visualizer/application/dtos/page_visualization_page_dto.py
@@ -1,7 +1,7 @@
 from typing import Optional, TypedDict
 
 
-class HtmlExportPageDto(TypedDict):
+class PageVisualizationPageDto(TypedDict):
     uid: str
     label: str
     name: str